+++ nczarr_test/zmapio.c
@@ -3,6 +3,7 @@
  *      See netcdf/COPYRIGHT file for copying and redistribution conditions.
  */
 
+#include <stddef.h>
 #ifdef HAVE_CONFIG_H
 #include "config.h"
 #endif
@@ -249,7 +250,7 @@ implfor(const char* path)
     NCURI* uri = NULL;
     const char* mode = NULL;
     NClist* segments = nclistnew();
-    int i;
+    size_t i;
     NCZM_IMPL impl = NCZM_UNDEF;
 
     ncuriparse(path,&uri);
@@ -318,7 +319,7 @@ objdump(void)
     NClist* stack = nclistnew();
     char* obj = NULL;
     char* content = NULL;
-    int depth;
+    size_t depth;
 
     if((stat=nczmap_open(dumpoptions.impl, dumpoptions.infile, NC_NOCLOBBER, 0, NULL, &map)))
         goto done;
@@ -327,10 +328,10 @@ objdump(void)
     if((stat = breadthfirst(map,"/",stack))) goto done;
 
     if(dumpoptions.debug) {
-	int i;
+	size_t i;
         fprintf(stderr,"stack:\n");
         for(i=0;i<nclistlength(stack);i++)
-            fprintf(stderr,"[%d] %s\n",i,(char*)nclistget(stack,i));
+            fprintf(stderr,"[%zu] %s\n",i,(char*)nclistget(stack,i));
     }    
     for(depth=0;depth < nclistlength(stack);depth++) {
         size64_t len = 0;
@@ -359,7 +360,7 @@ objdump(void)
 		if(kind == OK_CHUNK) {
 		    len = ceildiv(len,dumpoptions.nctype->typesize);
 		}
-                printf("[%d] %s : (%llu)",depth,obj,len);
+                printf("[%zu] %s : (%llu)",depth,obj,len);
 		if(kind == OK_CHUNK &&  dumpoptions.nctype->nctype != NC_STRING)
                     printf(" (%s)",dumpoptions.nctype->typename);
                 printf(" |");
@@ -378,10 +379,10 @@ objdump(void)
 		}
 	        printf("|\n");
 	    } else {
-	        printf("[%d] %s : (%llu) ||\n",depth,obj,len);
+	        printf("[%zu] %s : (%llu) ||\n",depth,obj,len);
 	    }
 	} else {
-	    printf("[%d] %s\n",depth,obj);
+	    printf("[%zu] %s\n",depth,obj);
 	}
     }
 done:
